+++ examples/basic/CHANGELOG.md
@@ -7,6 +7,5 @@
 - Updated dependencies
   - @computesdk/e2b@1.0.0
   - @computesdk/vercel@1.0.0
-  - @computesdk/cloudflare@1.0.0
+  - @computesdk/daytona@1.0.0
   - computesdk@1.0.0
-  - @computesdk/fly@1.0.0